Add a flag to get the watched repos
I added this option because I'm watching a few repositories that I have interest in the changes
or I want to contribute in the future. So with this option, I can follow the issues and pull requests
for this repositories I do not own or I'm not a contributor at the moment.